Started on it and couldn't pause.
Added click functionality to the canvas.
Update decided to be stubborn to work with the way I wanted it to (elements array) so decided to revert that bit to using the grid array instead.

Can now draw single cells instead of the whole thing. Obviously this draws over what was there before.